+++ src/shared/types/layout-toggle.ts
@@ -1,0 +1,7 @@
+export const TOGGLE = {
+  GRID: 'grid',
+  LIST: 'list',
+  WORD: 'word',
+} as const;
+
+export type LayoutToggleType = (typeof TOGGLE)[keyof typeof TOGGLE];
